Pass a command target to ActionMenuWidget
  * We intend to introduce a "delegate" object to handle the
various actions outside of the context of the UI data-binding, in
this case the load and save actions for files and MIDI controllers.

  * As a first step, we rework the "ActionMenuWidget" to accept a
delegate as part of its initialization, and refer to the various
actions directly there instead of using the "propagateCommand"
hack.

  * We pass the main window object as the delegate, propagating it
through the "MainWidget" class, leading to a somewhat odd case of
passing "self" twice to a constructor, but that will eventually be
changed.