Fold main window class logic into AlesisVSysexApplication
  * Remember how, eight commits ago, I split the application class
from the main window class?  Now it's time to re-unite them...

  * ... except not the same way that they were.  Most of the code
was either setup (which doesn't need to be stuck on a QMainWindow
subclass in the first place), or a bounce-method for behavior
(which we can just do from the application once it has the
requisite member variables set up).

  * Practical upshot: No more QMainWindow subclass, it's all done
via configuration.