Add logic to allow multiple main windows without running afoul of the GC
  * A disappointment more bitter than quinine tonic without gin:
After all of the work put in to split the application-specific
bits of UI logic away from the stock UI classes, a combination of
factors between whatever FFI is used by Python and the design of
Qt requires reintroducing a new QWhatever subclass.

  * The commentary explains what's going on adequately enough as
far as I'm concerned.  About the ONLY positive side that I can see
here is that we now have a main window class in place in case we
need to do any OTHER messing about because of how Qt fails to
accommodate a "stock UI class" style of programming.

  * I'd be angry, but it's really not worth the effort.  I'm
certainly not going to hate Qt for this.  But I AM disappointed.
This could have been better designed.

  * Also, not my first choice of class name, but the other was...
less tactful.