Do MIDI port discovery prior to creating MIDI device instances
  * Rather than attempting to match a suitable MIDI port during
device object initialization, add a class method to enumerate
suitable ports, do the matching on the UI side (allowing for, but
not yet implementing, graceful failure when there is no port and
user control when there are multiple ports), and just open the
selected port when initializing the device object.

  * This can still fail badly if someone removes the device at a
"bad time".  Attempting to fail gracefully under such conditions
would be nice, but I have no plans to work on it at this time.

  * Some of the (unused) parameters here are to establish the
interface that we are going to need moving forward, so that client
code will need less adjustment as we add functionality.