Use SYSEX update messages as the save file format
  * This is compatible with the windows editor for the V25, or
would be if we were producing suitable SYSEX messages for the V25.

  * It is not compatible with the windows editor for the VMini
(for some reason the VMini editor uses a different file format).

  * It is not compatible with the windows editor for the VI49, but
can be made so (the VI49 editor is clearly a copy/paste job from
the VI61 editor, and they missed a bit with the save file encoder).

  * This also means that a V25 or VMini save file is a valid SYSEX
message to be sent to set the configuration on said hardware via
other tools.  This will not work with the VI-series, however, as
they use a "slotted" configuration and therefore a different
access protocol.