Auto-select "sane" initial model on startup
  * Rather than hardcoding the use of the VMini model, first check
for connected devices and use a model compatible with the first
such device found.  If there aren't any connected devices, use an
arbitrary valid model (the "first" subclass of AlesisModel).

  * This is mostly trying for the "pleasant surprise" effect of
managing to do something sane given the available hardware when
the program starts, but may end up confusing or annoying some
people.  If necessary, we should be able to revert this easily
enough.