Initial rough-in for VI49 model class
  * These are all of the various bits, roughly as I understand
them, with a default configuration matching the windows editor.

  * Attempting to read a configuration from an actual VI49 does
not work, for a couple of reasons.  Attempting to write a
configuration to an actual VI49 is probably a bad idea.

  * Save and load will probably "work", but not be compatible with
save files from the windows editor, due to a change to the request
format and an oddity with the message size field.

  * The UI is unusable with this model, due to the sheer number of
switches (thirty-six, at least it's not the VI61's fourty-eight)
driving the vertical aspect of the edit window to approximately
twice the height of my display.

  * The emphasis in "rough-in", clearly, is on "rough".