Mark model classes for "slotted" configuration
  * The VI49 uses a "slotted" configuration, while the VMini and
V25 do not.  We can also expect the VI25 and VI61 to use slotted
configuration and the V49 and V61 to not.  The Vortex devices will
probably be slotted.

  * We could either mark each model explicitly as slotted or not,
or establish a default and then mark only the models that differ
from the default.  In this case, we declare the default to be not
slotted, and mark the VI49 as an exception.

  * And none of this actually has an effect yet, but now that it
exists we can take advantage of it.