Only evaluate BasicComponent _PARAMS not overridden during initialization
  * Lazyness, but with a semantic change behind it: If we evaluate
the defaults "early", and are enforcing ranges on IntValue classes,
using invalid values for the defaults always causes an error.  If
we evaluate the defaults "late" then we only get an error for an
invalid default if it is not overridden, which lends some strength
to the concept of an "invalid default".