Use "transpose" (relative to base note) instead of "base_note"
  * A bit of model cleanup: The "base note" for an octave is not
an accurate concept.  Instead, what we have is a transposition of
up to twelve semitones in either direction, encoded as the number
of semitones plus twelve.

  * We introduce a "delta" concept to IntValue for encoding
purposes, and use this in creating a Transpose class that uses
this delta and a new range (-12 to 12) for the actual value.

  * And then we tie everything together by using the new class in
the Keys model class, and updating the default initializers and
affected test cases.