Kludge VI61 upload into working, maybe
  * tl;dr: In the five-layer dependency sandwich that we have for
a MIDI interface, the middle layer decided not to support the
semantic that we require, so we kludge around it with a 2ms delay
to try and get a usable result.  RtMidi must die, python-rtmidi
must die, and mido must die.

  * In more detail, and reiterating the contents of the commentary
added in this commit, it appears that the VI61 (and presumably the
other VI-series devices) does not appreciate recieving SYSEX
messages split over USB packet boundaries, and possibly not even
more than one SYSEX message per USB packet.  This failure doesn't
always happen, the chances are affected by things such as the
overall system load, the total length of the configuration being
uploaded, and possibly even the phase of the moon.  The obvious
solution here is to wait until each SYSEX packet has been sent to
the MIDI device before queueing the next one, and the Linux kernel
and libalsa provide interfaces for doing exactly this, but RtMidi
doesn't provide this interface, and this prevents python-rtmidi
and mido from providing it.

  * Instead of doing the right thing, then, we rely on each SYSEX
message fitting within a single USB packet, and the USB packets
being sent every 1ms.  Because there's no guarantee of sub-1ms
timer resolution, we attempt to delay for 2ms, which should be
sufficient.

  * Of course, this is difficult to test.  My first attempt at
implementation worked perfectly... and then I found that I'd made
the edits in the wrong source tree and thus hadn't tested them,
I'd just hit that lucky combination of system load and whatnot to
get an apparent success.  Basically, it's easy to get a false
positive, and you can't really show that it's a true positive
without using a USB packet capture.

  * The real solution here is to switch from using mido to using
some other MIDI interface library that actually exposes the "drain"
API.  Something for the future, maybe.